+++ backend/server/constants.go
@@ -1,0 +1,9 @@
+package utility
+
+const DEVELOPMENT bool = true
+const BASE_URL = "http://192.168.99.100:1323/"
+const SPONSOR_URL = "api/v1/sponsors"
+const MAILING_URL = "api/v1/mailing"
+const FAQ_URL = "api/v1/faq"
+const AUTH_TOKEN = "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhZG1pbiI6dHJ1ZSwiZXhwIjoxNTkzMTI5NjAwLCJ6SUQiOiJ6NTEyMzQ1NiJ9.jYC2qlpzAKIMPFywQ6pWIV1qat_h7OrorJ-zQM5jDpg"
+var JWT_SECRET = []byte("temp_secret_until_proper_secrets_are_implemented")
