+++ backend/server/mailing/mailing_test.go
@@ -1,11 +1,15 @@
-package main
+package mailing
 
 import (
 	"fmt"
 	"net/http"
 	"net/url"
 	"testing"
+	
+	. "csesoc.unsw.edu.au/m/v2/server"
 )
+
+const mailingRequestUrl =  BASE_URL + MAILING_URL;
 
 func TestJoinMessages(t *testing.T) {
 	t.Run("Join enquiries", func(t *testing.T) {
@@ -79,23 +83,25 @@ func TestEnquirySuccessful(t *testing.T) {
 	}
 
 	t.Run("Handle successful sponsorship enquiry", func(t *testing.T) {
-		resp, err := http.PostForm("http://localhost:1323/api/v1/mailing/sponsorship", formCorrectData)
+		resp, err := http.PostForm(mailingRequestUrl + "/sponsorship", formCorrectData)
 		if err != nil {
 			t.Errorf("Could not perform POST request")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusAccepted)
+		AssertStatus(t, resp.StatusCode, http.StatusAccepted)
 	})
 
 	t.Run("Handle successful general enquiry", func(t *testing.T) {
-		resp, err := http.PostForm("http://localhost:1323/api/v1/mailing/general", formCorrectData)
+		resp, err := http.PostForm(mailingRequestUrl + "/general", formCorrectData)
 		if err != nil {
 			t.Errorf("Could not perform POST request")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusAccepted)
+		AssertStatus(t, resp.StatusCode, http.StatusAccepted)
 	})
 }
 
@@ -107,13 +113,14 @@ func TestEnquiryUnsuccessful(t *testing.T) {
 			"body":  {"I'd like to sponsor CSESoc"},
 		}
 
-		resp, err := http.PostForm("http://localhost:1323/api/v1/mailing/sponsorship", formIncorrectData)
+		resp, err := http.PostForm(mailingRequestUrl + "/sponsorship", formIncorrectData)
 		if err != nil {
 			t.Errorf("Could not perform POST request")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusBadRequest)
+		AssertStatus(t, resp.StatusCode, http.StatusBadRequest)
 	})
 
 	t.Run("Handle request with missing email", func(t *testing.T) {
@@ -123,13 +130,14 @@ func TestEnquiryUnsuccessful(t *testing.T) {
 			"body":  {"I'd like to sponsor CSESoc"},
 		}
 
-		resp, err := http.PostForm("http://localhost:1323/api/v1/mailing/sponsorship", formIncorrectData)
+		resp, err := http.PostForm(mailingRequestUrl + "/sponsorship", formIncorrectData)
 		if err != nil {
 			t.Errorf("Could not perform POST request")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusBadRequest)
+		AssertStatus(t, resp.StatusCode, http.StatusBadRequest)
 	})
 
 	t.Run("Handle request with no body", func(t *testing.T) {
@@ -139,13 +147,14 @@ func TestEnquiryUnsuccessful(t *testing.T) {
 			"body":  {""},
 		}
 
-		resp, err := http.PostForm("http://localhost:1323/api/v1/mailing/sponsorship", formIncorrectData)
+		resp, err := http.PostForm(mailingRequestUrl + "/sponsorship", formIncorrectData)
 		if err != nil {
 			t.Errorf("Could not perform POST request")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusBadRequest)
+		AssertStatus(t, resp.StatusCode, http.StatusBadRequest)
 	})
 
 	// http://softwaretesterfriend.com/manual-testing/valid-invalid-email-address-format-validation/
@@ -177,68 +186,73 @@ func TestEnquiryUnsuccessful(t *testing.T) {
 				"body":  {"I'd like to sponsor CSESoc"},
 			}
 
-			resp, err := http.PostForm("http://localhost:1323/api/v1/mailing/sponsorship", formIncorrectData)
+			resp, err := http.PostForm(mailingRequestUrl + "/sponsorship", formIncorrectData)
 			if err != nil {
 				t.Errorf("Could not perform POST request")
+				return;
 			}
 			defer resp.Body.Close()
 
-			assertStatus(t, resp.StatusCode, http.StatusBadRequest)
+			AssertStatus(t, resp.StatusCode, http.StatusBadRequest)
 		})
 	}
 }
 
 func TestFeedbackSuccessful(t *testing.T) {
 	t.Run("Feedback Successful", func(t *testing.T) {
-		resp, err := http.PostForm("http://localhost:1323/api/v1/mailing/feedback", url.Values{
+		resp, err := http.PostForm(mailingRequestUrl + "/feedback", url.Values{
 			"name":  {"John Smith"},
 			"email": {"johnsmith@gmail.com"},
 			"body":  {"feedback message"},
 		})
 		if err != nil {
 			t.Errorf("could not perform POST request")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusAccepted)
+		AssertStatus(t, resp.StatusCode, http.StatusAccepted)
 	})
 
 	t.Run("Feedback missing name, missing email", func(t *testing.T) {
-		resp, err := http.PostForm("http://localhost:1323/api/v1/mailing/feedback", url.Values{
+		resp, err := http.PostForm(mailingRequestUrl + "/feedback", url.Values{
 			"body": {"feedback message"},
 		})
 		if err != nil {
 			t.Errorf("could not perform POST request")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusAccepted)
+		AssertStatus(t, resp.StatusCode, http.StatusAccepted)
 	})
 }
 
 func TestFeedbackError(t *testing.T) {
 	t.Run("Feedback missing name, missing email", func(t *testing.T) {
-		resp, err := http.PostForm("http://localhost:1323/api/v1/mailing/feedback", url.Values{
+		resp, err := http.PostForm(mailingRequestUrl + "/feedback", url.Values{
 			"email": {"abcde"},
 			"body":  {"feedback message"},
 		})
 		if err != nil {
 			t.Errorf("could not perform POST request")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusBadRequest)
+		AssertStatus(t, resp.StatusCode, http.StatusBadRequest)
 	})
 
 	t.Run("Feedback missing body", func(t *testing.T) {
-		resp, err := http.PostForm("http://localhost:1323/api/v1/mailing/feedback", url.Values{
+		resp, err := http.PostForm(mailingRequestUrl + "/feedback", url.Values{
 			"name": {"John Smith"},
 		})
 		if err != nil {
 			t.Errorf("could not perform POST request")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusBadRequest)
+		AssertStatus(t, resp.StatusCode, http.StatusBadRequest)
 	})
 }
