+++ backend/server/sponsor/sponsor_test.go
@@ -1,4 +1,4 @@
-package main
+package sponsor
 
 import (
 	"encoding/json"
@@ -7,24 +7,26 @@ import (
 	"strconv"
 	"strings"
 	"testing"
+	. "csesoc.unsw.edu.au/m/v2/server"
 )
 
 const companyName = "Example"
 const companyLogo = "https://static.canva.com/static/images/canva_logo_100x100@2x.png"
 const companyTier = "2"
 const companyDetail = "Example"
-const getRequest = "http://localhost:1323/api/v1/sponsors/"
+const companyUrl = "www.google.com"
+const sponsorRequestUrl =  BASE_URL + SPONSOR_URL;
 
 func TestSponsor(t *testing.T) {
-
 	t.Run("Sponsor setup test", func(t *testing.T) {
-		resp, err := http.Get("http://localhost:1323/api/v1/sponsors")
+		resp, err := http.Get(sponsorRequestUrl)
 		if err != nil {
 			t.Errorf("Could not get perform request.")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusOK)
+		AssertStatus(t, resp.StatusCode, http.StatusOK)
 		var sponsors []*Sponsor
 		if err = json.NewDecoder(resp.Body).Decode(&sponsors); err != nil {
 			t.Errorf("Error parsing json response: %s", err)
@@ -35,13 +37,14 @@ func TestSponsor(t *testing.T) {
 	})
 
 	t.Run("Testing sponsor filtering", func(t *testing.T) {
-		resp, err := http.Get("http://localhost:1323/api/v1/sponsors?tier=2")
+		resp, err := http.Get(sponsorRequestUrl + "?tier=2")
 		if err != nil {
 			t.Errorf("Could not perform get sponsors request. Check connection.")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusOK)
+		AssertStatus(t, resp.StatusCode, http.StatusOK)
 	})
 
 	t.Run("New sponsor", func(t *testing.T) {
@@ -51,68 +54,68 @@ func TestSponsor(t *testing.T) {
 			"logo":   {companyLogo},
 			"tier":   {companyTier},
 			"detail": {companyDetail},
+			"url": 	  {companyUrl},
 		}
-		req, _ := http.NewRequest("POST", "http://localhost:1323/api/v1/sponsors", strings.NewReader(form.Encode()))
-		req.Header.Add("Authorization", "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhZG1pbiI6dHJ1ZSwiZXhwIjoxNTkzMTI5NjAwLCJ6SUQiOiJ6NTEyMzQ1NiJ9.jYC2qlpzAKIMPFywQ6pWIV1qat_h7OrorJ-zQM5jDpg")
+		req, _ := http.NewRequest("POST", sponsorRequestUrl, strings.NewReader(form.Encode()))
+		req.Header.Add("Authorization", AUTH_TOKEN)
 		req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
 		resp, err := client.Do(req)
-		// resp, err := http.PostForm("http://localhost:1323/api/v1/sponsors", url.Values{
-		// 	"name":   {companyName},
-		// 	"logo":   {companyLogo},
-		// 	"tier":   {companyTier},
-		// 	"detail": {companyDetail},
-		// })
 		if err != nil {
 			t.Errorf("Could not perform post sponsor request. Check connection.")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusCreated)
+		AssertStatus(t, resp.StatusCode, http.StatusCreated)
 	})
 
 	t.Run("Get newly created sponsor", func(t *testing.T) {
-		resp, err := http.Get(getRequest + companyName)
+		resp, err := http.Get(sponsorRequestUrl + "/" + companyName)
 		if err != nil {
 			t.Errorf("Could not perform get sponsor request. Check connection.")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusOK)
+		AssertStatus(t, resp.StatusCode, http.StatusOK)
 
 		var newSponsor *Sponsor
 		if err = json.NewDecoder(resp.Body).Decode(&newSponsor); err != nil {
 			t.Errorf("Error parsing json response: %s", err)
 		} else {
-			assertResponseBody(t, newSponsor.Name, companyName)
-			assertResponseBody(t, newSponsor.Logo, companyLogo)
-			assertResponseBody(t, strconv.Itoa(newSponsor.Tier), companyTier)
+			AssertResponseBody(t, newSponsor.Name, companyName)
+			AssertResponseBody(t, newSponsor.Logo, companyLogo)
+			AssertResponseBody(t, strconv.Itoa(newSponsor.Tier), companyTier)
 		}
 	})
 
 	t.Run("Delete newly created sponsor", func(t *testing.T) {
 		client := &http.Client{}
-		req, err := http.NewRequest("DELETE", getRequest+companyName, nil)
-		req.Header.Add("Authorization", "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhZG1pbiI6dHJ1ZSwiZXhwIjoxNTkzMTI5NjAwLCJ6SUQiOiJ6NTEyMzQ1NiJ9.jYC2qlpzAKIMPFywQ6pWIV1qat_h7OrorJ-zQM5jDpg")
+		req, err := http.NewRequest("DELETE", sponsorRequestUrl + "/" + companyName, nil)
+		req.Header.Add("Authorization", AUTH_TOKEN)
 		if err != nil {
 			t.Errorf("Could not create delete request for sponsor.")
+			return;
 		}
 		resp, err := client.Do(req)
 		if err != nil {
 			t.Errorf("Could not perform delete sponsor request. Check connection.")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusNoContent)
+		AssertStatus(t, resp.StatusCode, http.StatusNoContent)
 	})
 
 	t.Run("Check newly removed sponsor", func(t *testing.T) {
-		resp, err := http.Get(getRequest + companyName)
+		resp, err := http.Get(sponsorRequestUrl + "/" + companyName)
 		if err != nil {
 			t.Errorf("Could not perform get sponsor request. Check connection.")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusNotFound)
+		AssertStatus(t, resp.StatusCode, http.StatusNotFound)
 	})
 }
 
@@ -125,48 +128,52 @@ func TestSponsorError(t *testing.T) {
 			"tier":   {companyTier},
 			"detail": {companyDetail},
 		}
-		req, _ := http.NewRequest("POST", "http://localhost:1323/api/v1/sponsors", strings.NewReader(form.Encode()))
+		req, _ := http.NewRequest("POST", sponsorRequestUrl, strings.NewReader(form.Encode()))
 		req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-		req.Header.Add("Authorization", "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhZG1pbiI6dHJ1ZSwiZXhwIjoxNTkzMTI5NjAwLCJ6SUQiOiJ6NTEyMzQ1NiJ9.jYC2qlpzAKIMPFywQ6pWIV1qat_h7OrorJ-zQM5jDpg")
+		req.Header.Add("Authorization", AUTH_TOKEN)
 		req.PostForm = form
 		resp, err := client.Do(req)
 		if err != nil {
 			t.Errorf("Could not perform post sponsor request. Check connection.")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusCreated)
+		AssertStatus(t, resp.StatusCode, http.StatusCreated)
 
-		req, _ = http.NewRequest("POST", "http://localhost:1323/api/v1/sponsors", strings.NewReader(form.Encode()))
+		req, _ = http.NewRequest("POST", sponsorRequestUrl, strings.NewReader(form.Encode()))
 		req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-		req.Header.Add("Authorization", "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhZG1pbiI6dHJ1ZSwiZXhwIjoxNTkzMTI5NjAwLCJ6SUQiOiJ6NTEyMzQ1NiJ9.jYC2qlpzAKIMPFywQ6pWIV1qat_h7OrorJ-zQM5jDpg")
+		req.Header.Add("Authorization", AUTH_TOKEN)
 		req.PostForm = form
 		resp, err = client.Do(req)
 		if err != nil {
 			t.Errorf("Could not perform post sponsor request. Check connection.")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusConflict)
+		AssertStatus(t, resp.StatusCode, http.StatusConflict)
 
-		req, err = http.NewRequest("DELETE", getRequest+companyName, nil)
-		req.Header.Add("Authorization", "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhZG1pbiI6dHJ1ZSwiZXhwIjoxNTkzMTI5NjAwLCJ6SUQiOiJ6NTEyMzQ1NiJ9.jYC2qlpzAKIMPFywQ6pWIV1qat_h7OrorJ-zQM5jDpg")
+		req, err = http.NewRequest("DELETE", sponsorRequestUrl + "/" + companyName, nil)
+		req.Header.Add("Authorization", AUTH_TOKEN)
 		if err != nil {
 			t.Errorf("Could not create delete request for sponsor.")
+			return;
 		}
 		resp, err = client.Do(req)
 		if err != nil {
 			t.Errorf("Could not perform delete sponsor request. Check connection.")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusNoContent)
+		AssertStatus(t, resp.StatusCode, http.StatusNoContent)
 	})
 
 	t.Run("Missing parameters when creating", func(t *testing.T) {
 		client := &http.Client{}
-		req, _ := http.NewRequest("POST", "http://localhost:1323/api/v1/sponsors", nil)
-		req.Header.Add("Authorization", "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhZG1pbiI6dHJ1ZSwiZXhwIjoxNTkzMTI5NjAwLCJ6SUQiOiJ6NTEyMzQ1NiJ9.jYC2qlpzAKIMPFywQ6pWIV1qat_h7OrorJ-zQM5jDpg")
+		req, _ := http.NewRequest("POST", sponsorRequestUrl, nil)
+		req.Header.Add("Authorization", AUTH_TOKEN)
 		form := url.Values{
 			"name": {companyName},
 			"logo": {companyLogo},
@@ -175,33 +182,21 @@ func TestSponsorError(t *testing.T) {
 		resp, err := client.Do(req)
 		if err != nil {
 			t.Errorf("Could not perform post sponsor request. Check connection.")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusBadRequest)
+		AssertStatus(t, resp.StatusCode, http.StatusBadRequest)
 	})
 
 	t.Run("Get non existent sponsor", func(t *testing.T) {
-		resp, err := http.Get(getRequest + "nonexistent")
+		resp, err := http.Get(sponsorRequestUrl + "nonexistent")
 		if err != nil {
 			t.Errorf("Could not perform get sponsor request. Check connection.")
+			return;
 		}
 		defer resp.Body.Close()
 
-		assertStatus(t, resp.StatusCode, http.StatusNotFound)
+		AssertStatus(t, resp.StatusCode, http.StatusNotFound)
 	})
-}
-
-func assertStatus(t *testing.T, got, want int) {
-	t.Helper()
-	if got != want {
-		t.Errorf("got status %d, want %d", got, want)
-	}
-}
-
-func assertResponseBody(t *testing.T, got, want string) {
-	t.Helper()
-	if got != want {
-		t.Errorf("Response body is wrong, got %s, want %s", got, want)
-	}
 }
