+++ docker-compose.yml
@@ -11,11 +11,11 @@ services:
             - CHOKIDAR_USEPOLLING=true
     backend:
         image: backend
-        build: ./backend/server
+        build: ./backend
         ports:
             - '1323:1323'
         volumes:
-            - ./backend/server:/app/server
+            - ./backend:/app/server
         container_name: backend
     mongo:
         image: mongo:latest
